Code cleanup
1. Separate UIBoard into it's own file
2. Increase tile font size
3. Add moves helper function to Game
4. Simplify Bitboard::positions implementation

code cleanup
