+++ android/app/src/test/java/com/siyaman/tictactoe/BoardTest.kt
@@ -5,8 +5,6 @@ import org.junit.Assert.assertThrows
 import org.junit.Test
 
 /**
- * Example local unit test, which will execute on the development machine (host).
- *
  * See [testing documentation](http://d.android.com/tools/testing).
  */
 class BoardTest {
