+++ android/app/src/test/java/com/siyaman/tictactoe/GameTest.kt
@@ -4,8 +4,6 @@ import org.junit.Assert.*
 import org.junit.Test
 
 /**
- * Example local unit test, which will execute on the development machine (host).
- *
  * See [testing documentation](http://d.android.com/tools/testing).
  */
 class GameTest {
